+++ robot-command/src/main/java/org/obolibrary/robot/CommandLineInterface.java
@@ -42,6 +42,7 @@ public class CommandLineInterface {
     m.addCommand("report", new ReportCommand());
     m.addCommand("template", new TemplateCommand());
     m.addCommand("unmerge", new UnmergeCommand());
+    m.addCommand("validate", new ValidateCommand());
     m.addCommand("validate-profile", new ValidateProfileCommand());
     m.addCommand("verify", new VerifyCommand());
     return m;