+++ robot-core/src/main/java/org/obolibrary/robot/export/Table.java
@@ -1,5 +1,6 @@
 package org.obolibrary.robot.export;
 
+import com.google.common.collect.Sets;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonArray;
@@ -31,6 +32,8 @@ public class Table {
   private RendererType displayRenderer = null;
   private RendererType sortRenderer = null;
 
+  private static final Set<String> basicFormats = Sets.newHashSet("tsv", "csv", "json", "xlsx");
+
   /**
    * Init a new Table.
    *
@@ -43,11 +46,13 @@ public class Table {
     sortColumns = new ArrayList<>();
 
     // Set renderer types based on format
-    if (format.equalsIgnoreCase("tsv") || format.equalsIgnoreCase("csv")) {
+    if (format == null || basicFormats.contains(format.toLowerCase())) {
       displayRenderer = RendererType.OBJECT_RENDERER;
     } else if (format.equalsIgnoreCase("html")) {
       displayRenderer = RendererType.OBJECT_HTML_RENDERER;
       sortRenderer = RendererType.OBJECT_RENDERER;
+    } else {
+      // TODO - unknown format
     }
   }
 
@@ -83,11 +88,29 @@ public class Table {
     Sheet sheet = wb.getSheetAt(0);
     org.apache.poi.ss.usermodel.Row headerRow = sheet.createRow(0);
     int colIdx = 0;
+    Map<Integer, String> rules = new HashMap<>();
     for (Column c : columns) {
       String name = c.getDisplayName();
       Cell xlsxCell = headerRow.createCell(colIdx);
       xlsxCell.setCellValue(name);
+
+      String displayRule = c.getDisplayRule();
+      if (displayRule != null) {
+        rules.put(colIdx, displayRule);
+      }
       colIdx++;
+    }
+
+    // Maybe add rules
+    if (!rules.isEmpty()) {
+      org.apache.poi.ss.usermodel.Row rulesRow = sheet.createRow(sheet.getLastRowNum() + 1);
+      for (int idx = 0; idx <= colIdx; idx++) {
+        if (rules.containsKey(idx)) {
+          String rule = rules.get(idx);
+          Cell xlsxCell = rulesRow.createCell(idx);
+          xlsxCell.setCellValue(rule);
+        }
+      }
     }
 
     // Add rows
@@ -215,26 +238,81 @@ public class Table {
    * @return HTML string
    */
   public String toHTML(String split) {
+    return toHTML(split, true, false);
+  }
+
+  /**
+   * Render the Table as an HTML string.
+   *
+   * @param split character to split multiple cell values on
+   * @param standalone if true, include header
+   * @param includeJS if true and standalone, include JS script for tooltips
+   * @return HTML string
+   */
+  public String toHTML(String split, boolean standalone, boolean includeJS) {
     StringBuilder sb = new StringBuilder();
-    sb.append("<head>\n")
-        .append(
-            "\t<link rel=\"stylesheet\" href=\"https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css\">\n")
-        .append("</head>\n")
-        .append("<body>\n")
-        .append("<table class=\"table table-striped\">\n")
+    if (standalone) {
+      // Add opening tags, style, and maybe js scripts
+      sb.append("<head>\n")
+          .append("\t<link rel=\"stylesheet\" href=\"")
+          .append("https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css\">\n");
+      if (includeJS) {
+        sb.append("\t<script src=\"https://code.jquery.com/jquery-3.5.1.slim.min.js\"></script>\n")
+            .append(
+                "\t<script src=\"https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js\"></script>\n")
+            .append(
+                "\t<script src=\"https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js\"></script>\n");
+      }
+      sb.append("</head>\n").append("<body>\n");
+    }
+    // Table start
+    sb.append("<table class=\"table table-bordered table-striped\">\n")
+        .append("<thead class=\"bg-dark text-white header-row\">\n")
         .append("<tr>\n");
 
+    // Add column headers
+    Map<Integer, String> rules = new HashMap<>();
+    int colIdx = 0;
     for (Column c : columns) {
       sb.append("\t<th>").append(c.getDisplayName()).append("</th>\n");
+      String displayRule = c.getDisplayRule();
+      if (displayRule != null) {
+        rules.put(colIdx, displayRule);
+      }
+      colIdx++;
+    }
+    sb.append("</tr>\n").append("</thead>\n");
+
+    // Maybe add rules
+    if (!rules.isEmpty()) {
+      sb.append("<thead class=\"bg-secondary text-white\">\n").append("<tr>\n");
+      for (int idx = 0; idx < colIdx; idx++) {
+        if (rules.containsKey(idx)) {
+          sb.append("\t<th>").append(rules.get(idx)).append("</th>\n");
+        } else {
+          sb.append("\t<th></th>\n");
+        }
+      }
+      sb.append("</tr>\n").append("</thead>\n");
     }
 
-    sb.append("</tr>\n");
-
+    // Add all table rows
     for (Row row : rows) {
       sb.append(row.toHTML(columns, split));
     }
-    sb.append("</table>");
-    sb.append("</body>");
+    sb.append("</table>\n");
+
+    if (standalone) {
+      // Add closing tag and script to activate tooltips
+      sb.append("</body>\n");
+      if (includeJS) {
+        sb.append("<script>\n")
+            .append("\t$(function () {\n")
+            .append("\t\t$('[data-toggle=\"tooltip\"]').tooltip()\n")
+            .append("\t})")
+            .append("</script>\n");
+      }
+    }
     return sb.toString();
   }